auth/src/main/java/by/clevertec/auth/service/CustomUserDetailsService.java
package by.clevertec.auth.service;

import by.clevertec.aspect.ServiceAspectLogger;
import by.clevertec.auth.domain.User;
import by.clevertec.auth.security.CustomUserDetails;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CustomUserDetailsService implements UserDetailsService {

    private final UserService userService;

    @Override
    @ServiceAspectLogger
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userService.getUserByUsername(username);
        return new CustomUserDetails(user);
    }

}
